Refactor price calculation into smartPrice component
